Add project-wide documentation and move Spring18Application to proper location
