+++ spring18/src/main/java/com/sylconnexity/spring18/dbschema/ClickRepository.java
@@ -4,6 +4,9 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
+/**
+ * A repository for creating Click queries.
+ */
 public interface ClickRepository extends CrudRepository<Click, Long> {
 
 }
