+++ spring18/src/main/java/com/sylconnexity/spring18/dbschema/LinkRepository.java
@@ -4,7 +4,14 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
+/**
+ * A repository for creating Link queries.
+ */
 public interface LinkRepository extends CrudRepository<Link, Long> {
 
+    /**
+     * @param publisherID The ID of a publisher
+     * @return A list of Links with the given publisherID
+     */
     List<Link> findByPublisherID(Long publisherID);
 }
