+++ spring18/src/main/java/com/sylconnexity/spring18/dbschema/Merchant.java
@@ -5,6 +5,11 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
+/**
+ * A class defining the schema for Merchants. Merchants are uniquely identified by an automatically generated merchantID.
+ * <br><br>
+ * Merchants also have an associated merchantName.
+ */
 @Entity // This tells Hibernate to make a table out of this class
 public class Merchant {
     @Id
@@ -15,6 +20,10 @@ public class Merchant {
 
     protected Merchant() {}
 
+    /**
+     * Constructs a Merchant with an automatically generated, unique merchantID.
+     * @param merchantName The merchant name associated with where the SYL Link landed
+     */
     public Merchant(String merchantName) {
         this.merchantName = merchantName;
     }
