+++ spring18/src/main/java/com/sylconnexity/spring18/dbschema/MerchantRepository.java
@@ -4,6 +4,9 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
+/**
+ * A repository for creating Merchant queries.
+ */
 public interface MerchantRepository extends CrudRepository<Merchant, Long> {
 
 }
