+++ spring18/src/main/java/com/sylconnexity/spring18/dbschema/PublisherRepository.java
@@ -4,6 +4,9 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
+/**
+ * A repository for creating Publisher queries.
+ */
 public interface PublisherRepository extends CrudRepository<Publisher, Long> {
 
 }
