+++ spring18/src/main/java/com/sylconnexity/spring18/dbschema/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains the database schema classes of the SYL Link Journal system.
+ */
+package com.sylconnexity.spring18.dbschema;
