+++ spring18/src/main/java/com/sylconnexity/spring18/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains the classes used in the SYL Link Journal system.
+ */
+package com.sylconnexity.spring18;
