+++ spring18/src/main/java/com/sylconnexity/spring18/util/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Utility classes such as random class and primitive generators.
+ */
+package com.sylconnexity.spring18.util;
