+++ spring18/src/test/java/com/sylconnexity/spring18/Spring18ApplicationTests.java
@@ -5,6 +5,9 @@ import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
+/**
+ * Tests for the application.
+ */
 @RunWith(SpringRunner.class)
 @SpringBootTest
 public class Spring18ApplicationTests {
